added Test cases

+++ photo-album-manager/src/main/java/inst/an/photoalbummanager/controller/AlbumController.java
@@ -1,4 +1,4 @@
-package in.inst.photoalbummanager.controller;
+package inst.an.photoalbummanager.controller;
 
 import java.net.URI;
 import java.util.List;
@@ -15,9 +15,9 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
-import in.inst.photoalbummanager.beans.Album;
-import in.inst.photoalbummanager.repo.AlbumRepository;
-import in.inst.photoalbummanager.repo.UserRepository;
+import inst.an.photoalbummanager.beans.Album;
+import inst.an.photoalbummanager.repo.AlbumRepository;
+import inst.an.photoalbummanager.repo.UserRepository;
 
 @RestController
 @RequestMapping("/albums")
@@ -38,6 +38,16 @@ public class AlbumController {
 		return albumRepo.findByUserId(this.userRepo.findByName(username).getId());
 	}
 	
+	@RequestMapping(path="/{id:[0-9]+}", method = RequestMethod.GET)
+	public ResponseEntity<?>  readOne(@PathVariable Long id) {
+		Album album = this.albumRepo.findOne(id);
+		if(album == null)
+			return ResponseEntity.notFound().build();
+		if(!validate(album))
+			return ResponseEntity.status(403).body("Forbidden access to this album!");
+		return ResponseEntity.ok(albumRepo.findOne(id));
+	}
+	
 	@RequestMapping(method = RequestMethod.POST)
 	public ResponseEntity<?> create(@RequestBody Album input) {
 		Album result = this.albumRepo.save(input);
@@ -47,7 +57,7 @@ public class AlbumController {
 		return ResponseEntity.created(location).body(result);
 	}
 	
-	@RequestMapping(path="/{id}", method = RequestMethod.PUT)
+	@RequestMapping(path="/{id:[0-9]+}", method = RequestMethod.PUT)
 	public ResponseEntity<?> udpate(@PathVariable Long id, @RequestBody Album input) {
 		Album result = this.albumRepo.save(input);
 		URI location = ServletUriComponentsBuilder
@@ -56,9 +66,23 @@ public class AlbumController {
 		return ResponseEntity.created(location).body(result);
 	}
 	
-	@RequestMapping(path="/{id}", method = RequestMethod.DELETE)
+	@RequestMapping(path="/all", method = RequestMethod.GET)
+	public ResponseEntity<?> readAll() {
+		String username = ((User)SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUsername();
+		if(!username.equals("admin"))
+			return ResponseEntity.status(403).body("Access Denied!!");
+		return ResponseEntity.ok(albumRepo.findAll());
+	}
+	
+	@RequestMapping(path="/{id:[0-9]+}", method = RequestMethod.DELETE)
 	public ResponseEntity<?> delete(@PathVariable Long id) {
-		System.out.println("--> "+id);
+		System.out.println("delete --> "+id);
+		Album album = this.albumRepo.findOne(id);
+		if(album == null)
+			return ResponseEntity.notFound().build();
+		if(!validate(album))
+			return ResponseEntity.status(403).body("Forbidden access to this album!");
+		
 		try {
 			this.albumRepo.delete(id);
 		} catch (Exception e) {
@@ -70,8 +94,11 @@ public class AlbumController {
 		return ResponseEntity.ok("Album deleted successfully!");
 	}
 	
-	@RequestMapping(path="/all", method = RequestMethod.GET)
-	public List<Album> readAll() {
-		return albumRepo.findAll();
+	/**
+	 * 
+	 */
+	private boolean validate(Album album) {
+		String username = ((User)SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUsername();
+		return album.getUserId().equals(this.userRepo.findByName(username).getId());
 	}
 }
